Parallelize coloration.
If you have cores to spare, you can turn up GOMAXPROCS and take a
bunch of time out of colorizing the output (final phase).  Most of the
other work can't be as easily parallelized as I'd have to compute
overlap and such, but this is a fast way to shave some time off the
processing.